Pressure plate system implemented
added pressure plate functionality to the game, so if the player steps on top of the pressure plate it disables collision for the door making it possible to enter. also makes it so the door changes to green to let the player know it''s open. the player has 1 second to passs through before the door closes if they step off of the pressure plate. made it work with player and box tag.